+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - In `elements.getQuery()` endpoint the `element` parameter has been renamed to `elementType`.
 - In `elements.relations()` endpoint the `element` parameter has been renamed to `elementType`.
 - In `elements.full()` endpoint the `element` parameter has been renamed to `elementType`.
+- Type of `user_id` parameter in `changeset.get_query()` was changed from `str` to `int`.
 
 ### Removed
 - Support for `HTTP Basic authentication`: `username` and `password` parameters in `Api` class constructor.